echo/requirements.txt
Flask==3.0.3
Flask-SQLAlchemy==3.1.1
Flask-Login==0.6.3
Flask-WTF==1.2.1
Flask-Migrate==4.0.7
email-validator==2.2.0
pytest==8.2.2
selenium==4.22.0

openai==1.99.9

python-dotenv==1.0.1
